19-06-2016
The code now compiles.
But ....

Strangely the MPU chip fail to initialize.
It seems the code doesn't retrieve a correct I2C ID from the chip.
That's what I don't understand : this is the lib job and and I didn't
mod it.
Oddly with other Code (current LSOS or calibration sketch) it works ok
...

I did test it with the DIYino board and with my hilt: same results.
Well as always ... nothing goes easy ...

Ohh .... and it only compiles on Eclipse due to "multiple definition
errors" .... T_T
The linker thinks there is circular dependencies...